+++ styles/styles.css
@@ -71,6 +71,9 @@ body.appear {
 body.overflow-hidden {
   overflow: hidden;
 }
+body.overflowY-hidden {
+  overflow-y: hidden;
+}
 body .modal {
   width: 100%;
   height: 100%;
