+++ styles/styles.less
@@ -102,6 +102,10 @@ body {
 		overflow: hidden;
 	}
 
+	&.overflowY-hidden {
+		overflow-y: hidden;
+	}
+
 	.modal {
         width: 100%;
         height: 100%;
